ble: Organize ble classes and services
This patch separate what is ble implementations service and knot and
openthread implementations services.
